requirements.txt
streamlit
pymongo
python-dotenv
streamlit-option-menu